done ex 4.21

+++ part4/blogList/controllers/blogs.js
@@ -2,6 +2,7 @@ const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const { error } = require('../utils/logger')
 
 // const getTockenFrom = (request) => {
 //   const authorization = request.get('authorization')
@@ -40,7 +41,7 @@ blogsRouter.post('/', async (request, response) => {
   if (!decodedTocken.id) {
     return response
       .status(401)
-      .json({ error: 'tocken invalid' })
+      .json({ error: 'token invalid' })
   }
 
   const user = await User.findById(decodedTocken.id)
@@ -48,7 +49,9 @@ blogsRouter.post('/', async (request, response) => {
   // const user = await User.findById(body.userId)
 
   if (!user) {
-    return response.status(400).json({ error: 'userId missing or not valid' })
+    return response
+      .status(400)
+      .json({ error: 'userId missing or not valid' })
   }
   
   const blog = new Blog({
@@ -62,7 +65,9 @@ blogsRouter.post('/', async (request, response) => {
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
-  response.status(201).json(savedBlog)
+  response
+    .status(201)
+    .json(savedBlog)
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -84,6 +89,36 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedTocken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedTocken.id) {
+    return response
+      .status(401)
+      .json({ error: 'token invalid' })
+  }
+
+  // const user = await User.findById(decodedTocken.id)
+  // if (!user) {
+  //   return response
+  //     .status(400)
+  //     .json({ error: 'userId missing or not valid' })
+  // }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response
+      .status(404)
+      .json({ error: 'blog not found' })
+  }
+
+  console.log('blog.user.id', blog.user.toString())
+  console.log('decodedTocken.id', decodedTocken.id.toString())
+
+  if (blog.user.toString() !== decodedTocken.id.toString()) {
+    return response
+      .status(401)
+      .json({ error: 'only the creator can delete a blog' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
